--- src/main/java/sparta/spartateamproject1/exception/Advice.java
@@ -1,18 +1,0 @@
-package sparta.spartateamproject1.exception;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class Advice {
-    @ExceptionHandler(LoginException.class)
-    public ResponseEntity<?> handleLoginException(LoginException e) {
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
-
-    @ExceptionHandler(AdminNotFoundException.class)
-    public ResponseEntity<?> handleAdminNotFoundException(AdminNotFoundException e) {
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
-}
